backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain & LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.9
openai==1.6.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
asyncpg==0.29.0
alembic==1.13.1

# Document Parsing
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
uuid6==2024.1.12
aiofiles==23.2.1
